TPRS runs for 8 of the 26 ponds
Run these 8 through TPRS then merge with finished results dataframe bellow. Keep it seperate and run the same summay stats and make plots with these 8 ponds comparing all four models. Also include maps in the slides. I think the most likely outcome is that we decide to just go with TIN (in which case we are done, because we already have that calculated out for all of the ponds)